fix: fix in/de crease energy
Solve the bug with energy reduction when the airplane is pushed to the boundary and the product of velocity with vector is not equal to zero

Now there is an additional check for movement along one or another coordinate axis and a boolean variable showing whether there is movement in the current tick.

If there is motion and the speed is 2, it takes away energy, if there is no motion, it increases energy.